fix caching

+++ src/main/java/de/jensknipper/re_director/web/controller/dto/RedirectResponse.java
@@ -3,4 +3,4 @@ package de.jensknipper.re_director.web.controller.dto;
 import de.jensknipper.re_director.db.entity.Status;
 
 public record RedirectResponse(
-    long id, String source, String target, int httpStatusCode, Status status) {}
+    int id, String source, String target, int httpStatusCode, Status status) {}
